adding single instance to db complete. Just realized need two new tables to retain multiple credits and debits for same envelope.

+++ db/initDb.sql
@@ -13,9 +13,9 @@
 		, color_g int
 		, color_b int
 		, creditRecursionDate int
-		, creditRecursionAmount int
+		, creditRecursionAmount numeric(7,2)
 		, debitRecursionDate int
-		, debitRecursionAmount int
+		, debitRecursionAmount numeric(7,2)
 	);
 	INSERT INTO envelopes
     (
